Introduces incremental hash.
Fixes #181. Each API is subtly different, but in general we have a
new/update/finalize/reset.

The WebCrypto API for browsers doesn't support incremental hashes
yet; we could either shim it (i.e. build up the `ByteVector` in
memory) or just throw. I feel that it is better to throw so that the
user does not have an expectation of memory usage, but it might be a
PITA for anyone cross-compling.